+++ src/engine/Engine.ts
@@ -203,7 +203,7 @@ export class Engine extends Class implements CanInitialize, CanUpdate, CanDraw {
 
   /**
    * Direct access to the engine's 2D rendering context
-   * @deprecated will be removed v0.26.0
+   * @deprecated Will be removed in v0.26.0, use Engine.graphicsContext
    */
   public ctx: CanvasRenderingContext2D;
 
@@ -396,6 +396,17 @@ export class Engine extends Class implements CanInitialize, CanUpdate, CanDraw {
    * Sets the Transparency for the engine.
    */
   public enableCanvasTransparency: boolean = true;
+
+  /**
+   * Hints the graphics context to truncate fractional world space coordinates
+   */
+  public get snapToPixel(): boolean {
+    return this.graphicsContext.snapToPixel;
+  };
+
+  public set snapToPixel(shouldSnapToPixel: boolean) {
+    this.graphicsContext.snapToPixel = shouldSnapToPixel;
+  };
 
   /**
    * The action to take when a fatal exception is thrown
@@ -1208,7 +1219,6 @@ O|===|* >________________>\n\
       // Configure resolution for loader, it expects resolution === viewport
       this.screen.resolution = this.screen.viewport;
       this.screen.applyResolutionAndViewport();
-      this.graphicsContext.updateViewport();
       this._loader = loader;
       this._loader.suppressPlayButton = this._suppressPlayButton || this._loader.suppressPlayButton;
       this._loader.wireEngine(this);
@@ -1228,7 +1238,6 @@ O|===|* >________________>\n\
       // reset back to previous user resolution/viewport
       this.screen.popResolutionAndViewport();
       this.screen.applyResolutionAndViewport();
-      this.graphicsContext.updateViewport();
     }
 
     this._loadingComplete = true;