+++ src/engine/Scene.ts
@@ -349,9 +349,7 @@ export class Scene extends Class implements CanInitialize, CanActivate, CanDeact
    */
   public update(engine: Engine, delta: number) {
     this._preupdate(engine, delta);
-    if (this.camera) {
-      this.camera.update(engine, delta);
-    }
+
     // TODO differed entity removal for timers
     let i: number, len: number;
     // Remove timers in the cancel queue before updating them
@@ -366,6 +364,11 @@ export class Scene extends Class implements CanInitialize, CanActivate, CanDeact
     }
 
     this.world.update(SystemType.Update, delta);
+
+    // Camera last keeps renders smooth that are based on entity/actor
+    if (this.camera) {
+      this.camera.update(engine, delta);
+    }
 
     this._collectActorStats(engine);